put precedence logic on expression
