+++ src/generators/nodes/shared/Expression.ts
@@ -3,6 +3,55 @@ import { walk } from 'estree-walker';
 import isReference from 'is-reference';
 import flattenReference from '../../../utils/flattenReference';
 import { createScopes } from '../../../utils/annotateWithScopes';
+import { Node } from '../../../interfaces';
+
+const binaryOperators: Record<string, number> = {
+	'**': 15,
+	'*': 14,
+	'/': 14,
+	'%': 14,
+	'+': 13,
+	'-': 13,
+	'<<': 12,
+	'>>': 12,
+	'>>>': 12,
+	'<': 11,
+	'<=': 11,
+	'>': 11,
+	'>=': 11,
+	'in': 11,
+	'instanceof': 11,
+	'==': 10,
+	'!=': 10,
+	'===': 10,
+	'!==': 10,
+	'&': 9,
+	'^': 8,
+	'|': 7
+};
+
+const logicalOperators: Record<string, number> = {
+	'&&': 6,
+	'||': 5
+};
+
+const precedence: Record<string, (node?: Node) => number> = {
+	Literal: () => 21,
+	Identifier: () => 21,
+	ParenthesizedExpression: () => 20,
+	MemberExpression: () => 19,
+	NewExpression: () => 19, // can be 18 (if no args) but makes no practical difference
+	CallExpression: () => 19,
+	UpdateExpression: () => 17,
+	UnaryExpression: () => 16,
+	BinaryExpression: (node: Node) => binaryOperators[node.operator],
+	LogicalExpression: (node: Node) => logicalOperators[node.operator],
+	ConditionalExpression: () => 4,
+	AssignmentExpression: () => 3,
+	YieldExpression: () => 2,
+	SpreadElement: () => 1,
+	SequenceExpression: () => 0
+};
 
 export default class Expression {
 	compiler: Generator;
@@ -115,6 +164,10 @@ export default class Expression {
 
 		this.contexts = new Set(); // TODO...
 		this.indexes = new Set(); // TODO...
+	}
+
+	getPrecedence() {
+		return this.node.type in precedence ? precedence[this.node.type](this.node) : 0;
 	}
 
 	overwriteThis(name) {
